feat: move `extendsList` to `rules` folder + rename to `rulesExtends`
